requirements.txt
google-generativeai>=0.8.0
numpy>=1.25.0
scipy>=1.10.0
pyaudio>=0.2.14
pvporcupine>=3.0.0
python-dotenv>=1.0.0
webrtcvad>=2.0.10
